sensors/requirements.txt
schedule
grequests
influxdb-client